refactor: standardize request validation
